// src/app/interfaces/i_process-info-texts.ts
export interface IProcessInfoTexts {
    Title: string;

    CustomerDetails: string;
    Name: string;
    PersonalId: string;
    Age: string;
    Address: string;
    
    SuperClaimType: string
    EventDate: string;
    ClaimCause: string;
    InjuryType: string;
    SubmitedBy: string;
    SubmitionMethod: string;

    Select: string;
    SelectClaimCause: string;
}